Convert metabase/parameters/utils/ui to TypeScript (#24566)
